Implement authorization by policy

+++ app/Http/Controllers/QuestionController.php
@@ -8,6 +8,10 @@ use Illuminate\Http\Request;
 
 class QuestionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -55,7 +59,7 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        //
+        $this->authorize('update',$question);
         return view('questions.edit',['question'=>$question]);
     }
 
@@ -64,7 +68,7 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        //
+        $this->authorize('update',$question);
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with(['success'=>"Your question has been updated success"]);
     }
@@ -74,7 +78,7 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
-        //
+        $this->authorize('delete',$question);
         $question->delete();
         return redirect()->route('questions.index')->with(['success'=>"Your question has been deleted"]);
     }
